init: move init manager code
Needed for LDS. No actual logic changes just code movement.
