+++ source/server/server.h
@@ -23,6 +23,7 @@
 
 #include "server/connection_handler_impl.h"
 #include "server/http/admin.h"
+#include "server/init_manager_impl.h"
 #include "server/listener_manager_impl.h"
 #include "server/test_hooks.h"
 #include "server/worker.h"
@@ -77,24 +78,6 @@ public:
    * integration tests where a mock runtime is not needed.
    */
   static Runtime::LoaderPtr createRuntime(Instance& server, Server::Configuration::Initial& config);
-};
-
-/**
- * Implementation of Init::Manager for use during post cluster manager init / pre listening.
- */
-class InitManagerImpl : public Init::Manager {
-public:
-  void initialize(std::function<void()> callback);
-
-  // Init::Manager
-  void registerTarget(Init::Target& target) override;
-
-private:
-  enum class State { NotInitialized, Initializing, Initialized };
-
-  std::list<Init::Target*> targets_;
-  State state_{State::NotInitialized};
-  std::function<void()> callback_;
 };
 
 /**
